requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Deprecated==1.2.13
fastapi==0.92.0
h11==0.14.0
idna==3.4
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
pydantic==1.10.5
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
wrapt==1.14.1
